(fix) change strategy name from uniwap_v3 to amm_v3
