+++ Makefile
@@ -21,7 +21,7 @@ test:
  	--exclude-dir="test/hummingbot/connector/gateway/clob_spot/data_sources/dexalot" \
  	--exclude-dir="test/hummingbot/strategy/amm_arb" \
  	--exclude-dir="test/hummingbot/core/gateway" \
- 	--exclude-dir="test/hummingbot/strategy/uniswap_v3_lp"
+ 	--exclude-dir="test/hummingbot/strategy/amm_v3_lp"
 
 run_coverage: test
 	coverage report
