+++ hummingbot/strategy/amm_v3_lp/start.py
@@ -1,8 +1,8 @@
 from decimal import Decimal
 
+from hummingbot.strategy.amm_v3_lp.amm_v3_lp import AmmV3LpStrategy
+from hummingbot.strategy.amm_v3_lp.amm_v3_lp_config_map import amm_v3_lp_config_map as c_map
 from hummingbot.strategy.market_trading_pair_tuple import MarketTradingPairTuple
-from hummingbot.strategy.uniswap_v3_lp.uniswap_v3_lp import UniswapV3LpStrategy
-from hummingbot.strategy.uniswap_v3_lp.uniswap_v3_lp_config_map import uniswap_v3_lp_config_map as c_map
 
 
 def start(self):
@@ -18,8 +18,8 @@ def start(self):
 
     market_info = MarketTradingPairTuple(self.markets[connector], pair, base, quote)
     self.market_trading_pair_tuples = [market_info]
-    self.strategy = UniswapV3LpStrategy(market_info,
-                                        fee_tier,
-                                        price_spread,
-                                        amount,
-                                        min_profitability)
+    self.strategy = AmmV3LpStrategy(market_info,
+                                    fee_tier,
+                                    price_spread,
+                                    amount,
+                                    min_profitability)
